remove s3fs and smart_open from requirments
