+++ featuretools/entityset/deserialize.py
@@ -4,16 +4,12 @@ import tarfile
 import tempfile
 from pathlib import Path
 
-import boto3
 import pandas as pd
 
 from featuretools.entityset.relationship import Relationship
 from featuretools.entityset.serialize import FORMATS
-from featuretools.utils.gen_utils import (
-    check_schema_version,
-    use_s3fs_es,
-    use_smartopen_es
-)
+from featuretools.utils.gen_utils import check_schema_version
+from featuretools.utils.s3_utils import use_s3fs_es, use_smartopen_es
 from featuretools.utils.wrangle import _is_s3, _is_url
 from featuretools.variable_types.variable import LatLong, find_variable_types
 
@@ -190,6 +186,11 @@ def read_entityset(path, profile_name=None, **kwargs):
             kwargs (keywords): Additional keyword arguments to pass as keyword arguments to the underlying deserialization method.
     '''
     if _is_url(path) or _is_s3(path):
+        try:
+            import boto3
+        except ImportError:
+            raise ImportError("Please install boto3")
+
         with tempfile.TemporaryDirectory() as tmpdir:
             file_name = Path(path).name
             file_path = os.path.join(tmpdir, file_name)
