+++ featuretools/feature_base/features_deserializer.py
@@ -1,7 +1,5 @@
 import json
 
-import boto3
-
 from featuretools.entityset.deserialize import \
     description_to_entityset as deserialize_es
 from featuretools.feature_base.feature_base import (
@@ -15,11 +13,8 @@ from featuretools.feature_base.feature_base import (
     TransformFeature
 )
 from featuretools.primitives.utils import PrimitivesDeserializer
-from featuretools.utils.gen_utils import (
-    check_schema_version,
-    use_s3fs_features,
-    use_smartopen_features
-)
+from featuretools.utils.gen_utils import check_schema_version
+from featuretools.utils.s3_utils import use_s3fs_features, use_smartopen_features
 from featuretools.utils.wrangle import _is_s3, _is_url
 
 
@@ -93,6 +88,10 @@ class FeaturesDeserializer(object):
                 if _is_url(features):
                     features_dict = use_smartopen_features(features)
                 elif _is_s3(features):
+                    try:
+                        import boto3
+                    except ImportError:
+                        raise ImportError("Please install boto3")
                     session = boto3.Session()
                     if isinstance(profile_name, str):
                         transport_params = {'session': boto3.Session(profile_name=profile_name)}
