+++ requirements.txt
@@ -1,4 +1,3 @@
-python-dateutil>=2.6.1,<2.8.1
 scipy>=0.13.3
 numpy>=1.13.3
 pandas>=0.24.1
@@ -10,5 +9,3 @@ dask>=1.1.0
 psutil>=5.4.8
 click>=7.0.0
 scikit-learn>=0.20.0
-smart-open>=1.8.4
-s3fs >= 0.2.2
